--- a/Library/Homebrew/cmd/search.rb
+++ b/Library/Homebrew/cmd/search.rb
@@ -86,7 +86,8 @@ module Homebrew
     elsif args.pull_request?
       search_pull_requests(query, args)
     else
-      search_names(query, string_or_regex, args)
+      formulae, casks = search_names(query, string_or_regex)
+      print_results(formulae, casks, query, args)
     end
 
     puts "Use `brew desc` to list packages with a short description." if args.verbose?
@@ -127,5 +128,49 @@ module Homebrew
     end
 
     GitHub.print_pull_requests_matching(query, only)
+  end
+
+  def print_results(all_formulae, all_casks, query, args)
+    print_formulae = args.formula?
+    print_casks = args.cask?
+    print_formulae = print_casks = true if !print_formulae && !print_casks
+    print_formulae &&= all_formulae.any?
+    print_casks &&= all_casks.any?
+
+    count = 0
+    if all_formulae
+      if $stdout.tty?
+        ohai "Formulae", Formatter.columns(all_formulae)
+      else
+        puts all_formulae
+      end
+      count += all_formulae.count
+    end
+    puts if print_formulae && print_casks
+    if print_casks
+      if $stdout.tty?
+        ohai "Casks", Formatter.columns(all_casks)
+      else
+        puts all_casks
+      end
+      count += all_casks.count
+    end
+
+    print_missing_formula_help(query, count.positive?) if all_casks.exclude?(query)
+
+    odie "No formulae or casks found for #{query.inspect}." if count.zero?
+  end
+
+  def print_missing_formula_help(query, found_matches)
+    return unless $stdout.tty?
+
+    reason = MissingFormula.reason(query, silent: true)
+    return if reason.nil?
+
+    if found_matches
+      puts
+      puts "If you meant #{query.inspect} specifically:"
+    end
+    puts reason
   end
 end