--- a/Library/Homebrew/cmd/search.rb
+++ b/Library/Homebrew/cmd/search.rb
@@ -86,8 +86,8 @@ module Homebrew
     elsif args.pull_request?
       search_pull_requests(query, args)
     else
-      formulae, casks = search_names(query, string_or_regex)
-      print_results(formulae, casks, query, args)
+      formulae, casks = search_names(query, string_or_regex, args)
+      print_results(formulae, casks, query)
     end
 
     puts "Use `brew desc` to list packages with a short description." if args.verbose?
@@ -130,30 +130,23 @@ module Homebrew
     GitHub.print_pull_requests_matching(query, only)
   end
 
-  def print_results(all_formulae, all_casks, query, args)
-    print_formulae = args.formula?
-    print_casks = args.cask?
-    print_formulae = print_casks = true if !print_formulae && !print_casks
-    print_formulae &&= all_formulae.any?
-    print_casks &&= all_casks.any?
+  def print_results(all_formulae, all_casks, query)
+    count = all_formulae.size + all_casks.size
 
-    count = 0
-    if all_formulae
+    if all_formulae.any?
       if $stdout.tty?
         ohai "Formulae", Formatter.columns(all_formulae)
       else
         puts all_formulae
       end
-      count += all_formulae.count
     end
-    puts if print_formulae && print_casks
-    if print_casks
+    puts if all_formulae.any? && all_casks.any?
+    if all_casks.any?
       if $stdout.tty?
         ohai "Casks", Formatter.columns(all_casks)
       else
         puts all_casks
       end
-      count += all_casks.count
     end
 
     print_missing_formula_help(query, count.positive?) if all_casks.exclude?(query)