- [x] Have you followed the guidelines in our [Contributing](https://github.com/Homebrew/brew/blob/HEAD/CONTRIBUTING.md) document?
- [x] Have you checked to ensure there aren't other open [Pull Requests](https://github.com/Homebrew/brew/pulls) for the same change?
- [x] Have you added an explanation of what your changes do and why you'd like us to include them?
- [ ] Have you written new tests for your changes? [Here's an example](https://github.com/Homebrew/brew/blob/HEAD/Library/Homebrew/test/PATH_spec.rb).
- [x] Have you successfully run `brew style` with your changes locally?
- [x] Have you successfully run `brew typecheck` with your changes locally?
- [x] Have you successfully run `brew tests` with your changes locally?

-----

This closes #14183. Essentially, it uses `Search.search_names` to print suggestions in `brew install` if the cask/formula doesn't exist. This shouldn't actually be a lot more work because we were already hitting the Github API and getting the cask and formula info before. This just means that we'll display that info to the user.

Changes:
- Move `search_names` and `print_missing_formula_help` out of `cmd/search.rb` to `search.rb`
- Change to using those functions in `cmd/install.rb` when a formula or cask doesn't exist

Examples:
```
$ brew install -n --formula virtuabox
Warning: No available formula with the name "virtuabox". Did you mean virtuoso or virtualpg?
==> Searching for similarly named formulae and casks...
==> Formulae
virtuoso                                                                        virtualpg

To install one of them, run:
  brew install 'package_name'
```
```
$ brew install -n --cask virtuabox
Warning: Cask 'virtuabox' is unavailable: No Cask with this name exists.
==> Searching for similarly named formulae and casks...
==> Casks
virtualbox ✔                                                                    virtualgl

To install one of them, run:
  brew install 'package_name'
```
```
$ brew install -n virtuabox
Warning: No available formula with the name "virtuabox". Did you mean virtuoso or virtualpg?
==> Searching for similarly named formulae and casks...
==> Formulae
virtuoso                                                                        virtualpg

==> Casks
virtualbox ✔                                                                    virtualgl

To install one of them, run:
  brew install 'package_name'
```
```
$ brew install -n not_a_package
Warning: No available formula with the name "not_a_package".
==> Searching for similarly named formulae and casks...
Error: No formulae or casks found for "not_a_package".
```