@apainintheneck 2022-12-10
Also, there's a part of me that wants to make `search.rb` a module function instead of mixing it in with modules since it's included in a few different places. I think it'd make the code easier to reason about. That being said I decided it really wasn't necessary for this PR so I decided not to include it here.