@apainintheneck 2022-12-13
Now it looks like this.

```
$brew install -n google
Warning: No available formula with the name "google". Did you mean googler or goose?
==> Searching for similarly named formulae and casks...
==> Formulae
aws-google-auth                                       google-benchmark                                      google-java-format                                    google-sql-tool                                       googletest
google-authenticator-libpam                           google-go                                             google-sparsehash                                     googler                                               goose

To install aws-google-auth, run:
  brew install aws-google-auth

==> Casks
google-ads-editor                            google-chat                                  google-cloud-sdk                             google-earth-pro                             google-web-designer                          moefe-google-translate
google-analytics-opt-out                     google-chat-electron                         google-drive                                 google-japanese-ime                          googleappengine
google-assistant                             google-chrome                                google-drive-file-stream                     google-trends                                marshallofsound-google-play-music-player

To install google-ads-editor, run:
  brew install --cask google-ads-editor
```

I'm wondering if it might be a good idea to limit the number of search results. We don't need the search results to be complete or thorough in this command.

```
$brew install -n git-utils
Warning: No available formula with the name "git-utils". Did you mean eot-utils?
git-utils was deleted from homebrew/core in commit f29c64de4a1:
  git-utils: remove formula

To show the formula before removal, run:
  git -C "$(brew --repo homebrew/core)" show f29c64de4a1^:Formula/git-utils.rb

If you still use this formula, consider creating your own tap:
  https://docs.brew.sh/How-to-Create-and-Maintain-a-Tap
```

```
$ brew install -n not_a_tap/name
Warning: No available formula with the name "not_a_tap/name".
```