@MikeMcQuaid 2022-12-13
> To install aws-google-auth, run:
>  brew install aws-google-auth

Maybe want to make this have `--formula` to be consistent with `--cask` below?

> I'm wondering if it might be a good idea to limit the number of search results. We don't need the search results to be complete or thorough in this command.

Maybe? Can you produce any circumstances in which the unlimited version seems silly? If not, just leave it as-is (or for future PR at least).