@apainintheneck 2022-12-14
> > To install aws-google-auth, run:
> > brew install aws-google-auth
> 
> Maybe want to make this have `--formula` to be consistent with `--cask` below?

I mean it's not functionally necessary so... Plus, it'll make more sense without it for Linux users.

> > I'm wondering if it might be a good idea to limit the number of search results. We don't need the search results to be complete or thorough in this command.
> 
> Maybe? Can you produce any circumstances in which the unlimited version seems silly? If not, just leave it as-is (or for future PR at least).

The worst case is probably searching for something generic but it's not the end of the world. Let's show everything until people complain about it.

```
$ brew install -n google
Warning: No available formula with the name "google". Did you mean googler or goose?
==> Searching for similarly named formulae and casks...
==> Formulae
aws-google-auth                   google-java-format                googletest
google-authenticator-libpam       google-sparsehash                 goose
google-benchmark                  google-sql-tool
google-go ✔                       googler

To install aws-google-auth, run:
  brew install aws-google-auth

==> Casks
google-ads-editor                                  google-drive
google-analytics-opt-out                           google-drive-file-stream
google-assistant                                   google-earth-pro
google-chat                                        google-japanese-ime
google-chat-electron                               google-japanese-ime-dev
google-chrome                                      google-trends
google-chrome-beta                                 google-web-designer
google-chrome-canary                               googleappengine
google-chrome-dev                                  marshallofsound-google-play-music-player
google-cloud-sdk                                   moefe-google-translate

To install google-ads-editor, run:
  brew install --cask google-ads-editor
```