@apainintheneck 2022-12-14
Also, the cask search is kind of slow in comparison with formulae and hitting the Github API. I think I figured out why and I'll handle that in another PR.