@apainintheneck 2022-12-14
#### On this branch

```
/u/l/H/L/Homebrew (add-casks-to-install-search|✔) $ hyperfine 'brew search --cask google'
Benchmark 1: brew search --cask google
  Time (mean ± σ):      2.970 s ±  0.066 s    [User: 1.227 s, System: 1.072 s]
  Range (min … max):    2.902 s …  3.099 s    10 runs
 ```

#### On master
```
/u/l/H/L/Homebrew (master|✔) $ hyperfine 'brew search --cask google'
Benchmark 1: brew search --cask google
  Time (mean ± σ):      6.687 s ±  0.097 s    [User: 3.751 s, System: 1.903 s]
  Range (min … max):    6.586 s …  6.907 s    10 runs
```

This is on my old iMac but you get the idea.

https://github.com/Homebrew/brew/blob/ed3789a0bb34464a37acd224f196f906f8d73b6c/Library/Homebrew/extend/os/mac/search.rb#L40

Apparently, this line was very slow.