@MikeMcQuaid 2022-12-14
> I mean it's not functionally necessary so... Plus, it'll make more sense without it for Linux users.

Sure but: neither is `--cask`? Note: this feedback is not blocking.

> The worst case is probably searching for something generic but it's not the end of the world. Let's show everything until people complain about it.

👍🏻 

> Apparently, this line was very slow.

Oof! Nice work!