@apainintheneck 2022-12-15
> > I mean it's not functionally necessary so... Plus, it'll make more sense without it for Linux users.
> 
> Sure but: neither is `--cask`? Note: this feedback is not blocking.

I thought if there was any naming conflict that it would default to the formula. Is that not true?