@MikeMcQuaid 2022-12-15
> I thought if there was any naming conflict that it would default to the formula. Is that not true?

It would but it'll print a warning message which is a bit yucky.